refactor(groundcover): move GroundcoverIntegrationConfig to groundcover.config
